Ice Refraction
Created Ice refraction shader.

Can change the refraction via slider as well as the fresnell effect. Can also change color, but don¨t think that is gonna be needed.

OBS: shader will say that the texture should be changed to normal, but this will just "smooth" the material and "ignore" the texture. I assume it has something to do with it already being a normal map, and chaning it again alters some stuff, but i am not sure. Nevertheless, don't change the texture to normal and ignore unity's "suggestion"