Simple Buoyancy Script
a simple script to create floating object on still water by setting a specific height float in the inspector. In reality this would then automatically update to fit the moving water, but that would need to either calculate the entire math again or do an async callback to the shader based on the position.